Advanced task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	int x = 0, y = 1, z, w;
+	long int x = 0, y = 1, z, w;
 
 	for (w = 0; w < 50; w++)
 	{
@@ -16,9 +16,9 @@ int main(void)
 		x = y;
 		y = z;
 		if (w < 49)
-			printf("%d, ", z);
+			printf("%li, ", z);
 		else
-			printf("%d\n", z);
+			printf("%li\n", z);
 	}
 	return (0);
 }
